query-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
anthropic==0.40.0
mysql-connector-python==8.2.0
